Added multiplayer Functionality
It works by having a host player who has multplayer control and the other player is the peer. I havent found a way to make two diffferent players work on the same computer, however i think this is a problem with MAC computers not being able to have two tabs open at once.